+++ neo/Prompt/Commands/tests/test_config_commands.py
@@ -2,8 +2,8 @@ import os
 from neo.Settings import settings
 from neo.Utils.BlockchainFixtureTestCase import BlockchainFixtureTestCase
 from neo.Prompt.Commands.Config import CommandConfig
-from copy import deepcopy
-from neo.Network.NodeLeader import NodeLeader
+from neo.Network.NodeLeader import NodeLeader, NeoNode
+from neo.Network.address import Address
 from mock import patch
 from io import StringIO
 from neo.Prompt.PromptPrinter import pp
@@ -170,20 +170,64 @@ class CommandConfigTestCase(BlockchainFixtureTestCase):
             self.assertIn(f"Maintaining maxpeers at {settings.CONNECTED_PEER_MAX}", mock_print.getvalue())
 
         # test changing the number of maxpeers
-        args = ['maxpeers', "6"]
-        res = CommandConfig().execute(args)
-        self.assertTrue(res)
-        self.assertEqual(int(res), settings.CONNECTED_PEER_MAX)
+        with patch('sys.stdout', new=StringIO()) as mock_print:
+            args = ['maxpeers', "6"]
+            res = CommandConfig().execute(args)
+            self.assertTrue(res)
+            self.assertEqual(int(res), settings.CONNECTED_PEER_MAX)
+            self.assertIn(f"Maxpeers set to {settings.CONNECTED_PEER_MAX}", mock_print.getvalue())
 
         # test bad input
-        args = ['maxpeers', "blah"]
-        res = CommandConfig().execute(args)
-        self.assertFalse(res)
+        with patch('sys.stdout', new=StringIO()) as mock_print:
+            args = ['maxpeers', "blah"]
+            res = CommandConfig().execute(args)
+            self.assertFalse(res)
+            self.assertIn("Please supply a positive integer for maxpeers", mock_print.getvalue())
 
         # test negative number
-        args = ['maxpeers', "-1"]
-        res = CommandConfig().execute(args)
-        self.assertFalse(res)
+        with patch('sys.stdout', new=StringIO()) as mock_print:
+            args = ['maxpeers', "-1"]
+            res = CommandConfig().execute(args)
+            self.assertFalse(res)
+            self.assertIn("Please supply a positive integer for maxpeers", mock_print.getvalue())
+
+        # test if the new maxpeers < settings.CONNECTED_PEER_MAX
+        # first make sure we have a predictable state
+        NodeLeader.Instance().Reset()
+        leader = NodeLeader.Instance()
+        addr1 = Address("127.0.0.1:20333")
+        addr2 = Address("127.0.0.1:20334")
+        leader.ADDRS = [addr1, addr2]
+        leader.DEAD_ADDRS = [Address("127.0.0.1:20335")]
+        test_node = NeoNode()
+        test_node.host = "127.0.0.1"
+        test_node.port = 20333
+        test_node.address = Address("127.0.0.1:20333")
+        test_node2 = NeoNode()
+        test_node2.host = "127.0.0.1"
+        test_node2.port = 20333
+        test_node2.address = Address("127.0.0.1:20334")
+        leader.Peers = [test_node, test_node2]
+
+        with patch("neo.Network.NeoNode.NeoNode.Disconnect") as mock_disconnect:
+            # first test if the number of connected peers !< new maxpeers
+            with patch('sys.stdout', new=StringIO()) as mock_print:
+                args = ['maxpeers', "4"]
+                res = CommandConfig().execute(args)
+                self.assertTrue(res)
+                self.assertEqual(len(leader.Peers), 2)
+                self.assertFalse(mock_disconnect.called)
+                self.assertIn(f"Maxpeers set to {settings.CONNECTED_PEER_MAX}", mock_print.getvalue())
+
+            # now test if the number of connected peers < new maxpeers
+            with patch('sys.stdout', new=StringIO()) as mock_print:
+                args = ['maxpeers', "1"]
+                res = CommandConfig().execute(args)
+                self.assertTrue(res)
+                self.assertEqual(len(leader.Peers), 1)
+                self.assertEqual(leader.Peers[0].address, test_node.address)
+                self.assertTrue(mock_disconnect.called)
+                self.assertIn(f"Maxpeers set to {settings.CONNECTED_PEER_MAX}", mock_print.getvalue())
 
     def test_config_nep8(self):
         # test with missing flag argument